Add Newton l10n data abstract
